Fix integration tests

+++ reptor/plugins/tools/Zap/tests/test_integration_zap.py
@@ -14,7 +14,8 @@ class TestIntegrationZap(object):
         yield
         # Delete Zap notes (prevents interference between xml and json)
         note = notes_api.get_note_by_title("Zap")
-        notes_api.delete_note(note.id)
+        if note:
+            notes_api.delete_note(note.id)
 
     @pytest.mark.parametrize("format", ["xml", "json"])
     def test_notes(self, format, notes_api):  # noqa: F811
@@ -73,6 +74,6 @@ class TestIntegrationZap(object):
 
         for finding in findings:
             if finding.data.title == "Cross Site Scripting (Reflected)":
-                assert finding.data.severity.value == "critical"
+                assert finding.data.severity == "critical"
             elif finding.data.title == "Application Error Disclosure":
-                assert finding.data.severity.value == "high"
+                assert finding.data.severity == "high"
